matchcode/requirements.txt
asgiref==3.5.2
attrs==22.1.0
banal==1.0.6
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.6.0
boolean.py==4.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commoncode==31.0.0
container-inspector==32.0.1
cryptography==38.0.4
debian-inspector==31.0.0
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
dockerfile-parse==1.2.0
dparse2==0.6.1
extractcode==31.0.0
extractcode-7z==16.5.210531
extractcode-libarchive==3.5.1.210531
fasteners==0.18
fingerprints==1.0.3
ftfy==6.1.1
future==0.18.2
gemfileparser2==0.9.3
html5lib==1.1
idna==3.4
importlib-metadata==5.1.0
intbitset==3.0.1
isodate==0.6.1
jaraco.functools==3.5.2
javaproperties==0.8.1
Jinja2==3.1.2
jsonstreams==0.6.0
license-expression==30.0.0
lxml==4.9.1
MarkupSafe==2.1.1
more-itertools==9.0.0
natsort==8.2.0
normality==2.4.0
packagedb==2.0.0
packageurl-python==0.10.4
packaging==21.3
parameter-expansion-patched==0.3.1
pdfminer.six==20221105
pefile==2022.5.30
pip==21.1.2
pip-requirements-parser==31.2.0
pkginfo2==30.0.0
pluggy==1.0.0
plugincode==31.0.0
ply==3.11
psycopg2==2.9.3
psycopg2-binary==2.9.5
publicsuffix2==2.20191221
pyahocorasick==2.0.0b1
pycparser==2.21
pygmars==0.8.0
Pygments==2.13.0
pymaven-patch==0.3.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
rdflib==6.2.0
requests==2.28.1
saneyaml==0.5.2
scancode-toolkit @ https://github.com/nexB/scancode-toolkit/archive/refs/heads/maven-pom-parse-dep-backport.zip
setuptools==57.0.0
six==1.16.0
soupsieve==2.3.2.post1
spdx-tools==0.7.0a3
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
typecode==30.0.0
typecode-libmagic==5.39.210531
urllib3==1.26.13
urlpy==0.5
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
xmltodict==0.13.0
zipp==3.11.0
